pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.47.0"
torch = "^2.5.1"
pillow = "^11.0.0"
openai = "^1.57.4"
flask = "^3.1.0"
python-dotenv = "^1.0.1"


[[tool.poetry.source]]
name = "fpho"
url = "https://files.pythonhosted.org"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
